+++ NEWS.md
@@ -1,3 +1,5 @@
+# dbplyr (development version)
+
 # dbplyr 2.1.0
 
 ## New features