+++ NEWS.md
@@ -1,5 +1,8 @@
 # dbplyr (development version)
 
+* `compute()`, `sql_table_index()`, and `sql_query_wrap()` now work with
+  schemas (@mgirlich, #595).
+
 # dbplyr 2.1.0
 
 ## New features