+++ NEWS.md
@@ -1,6 +1,6 @@
 # dbplyr (development version)
 
-* `group_by()` now ungroups when the dots argument is empty and `.add` is `TRUE`
+* `group_by()` now ungroups when the dots argument is empty and `.add` is `FALSE`
   (@mgirlich, #615).
 
 * `compute()`, `sql_table_index()`, and `sql_query_wrap()` now work with