+++ NEWS.md
@@ -1,6 +1,12 @@
 # dbplyr (development version)
 
+
 * Added `sql_escape_date.ACCESS()` and `sql_escape_datetime.ACCESS`  to properly escape dates when using MS Access (@erikvona, #608).
+
+* Adds support for Snowflake databases (@edgararuiz)
+
+* `group_by()` now ungroups when the dots argument is empty and `.add` is `FALSE`
+  (@mgirlich, #615).
 
 * `compute()`, `sql_table_index()`, and `sql_query_wrap()` now work with
   schemas (@mgirlich, #595).