+++ NEWS.md
@@ -1,4 +1,4 @@
-# dbplyr (development version)
+# dbplyr 2.1.1
 
 * New support for Snowflake (@edgararuiz)